move categoryinput from component into parent
