--- components/form-category-input/category-input.styled.js
@@ -1,11 +1,0 @@
-import styled from "styled-components";
-export const StyledCategoryView = styled.div`
-  width: 100%;
-  height: 3em;
-  border: 2px solid #2490b5;
-  > * {
-    color: whitesmoke;
-    font-weight: bold;
-    position: relative;
-  }
-`;
